tests/wpt/web-platform-tests/portals/portals-cross-origin-load.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  promise_test(async () => {
    assert_implements("HTMLPortalElement" in self);
    var portal = document.createElement("portal");
    portal.src = "http://{{hosts[alt][www]}}:{{ports[http][0]}}/portals/resources/simple-portal.html";
    return new Promise((resolve, reject) => {
      portal.onload = resolve;
      document.body.appendChild(portal);
    });
  });
</script>
</body>
